AI_DRIVE_HELPER-main/requirements.txt
fastapi
uvicorn
python-multipart

pydantic
pydantic-settings

sqlalchemy
psycopg2-binary

PyJWT
bcrypt

google-api-python-client
google-auth

supabase
qdrant-client

insightface
onnxruntime
numpy
Pillow

streamlit
requests